Change mesh parsing to use file_t instead of FILE.
Since we want to use the BIN file instead of the ISO, we can just know
the start sector and read until we are finished. We have to use the
new functions in bin.h/c to read entire files into a buffer and then
parse them, maintaining our own internal offset/pointer.